# perf(831): Remove row_pk computation from query