--- a/crates/core/src/subscription/query.rs
+++ b/crates/core/src/subscription/query.rs
@@ -214,7 +214,6 @@ mod tests {
     use spacetimedb_lib::error::ResultTest;
     use spacetimedb_lib::Identity;
     use spacetimedb_primitives::{ColId, TableId};
-    use spacetimedb_sats::data_key::ToDataKey;
     use spacetimedb_sats::db::auth::{StAccess, StTableType};
     use spacetimedb_sats::db::def::*;
     use spacetimedb_sats::relation::FieldName;
@@ -223,28 +222,18 @@ mod tests {
     use spacetimedb_vm::operator::OpCmp;
 
     fn insert_op(table_id: TableId, table_name: &str, row: ProductValue) -> DatabaseTableUpdate {
-        let row_pk = row.to_data_key().to_bytes();
         DatabaseTableUpdate {
             table_id,
             table_name: table_name.to_string(),
-            ops: vec![TableOp {
-                op_type: 1,
-                row,
-                row_pk,
-            }],
+            ops: vec![TableOp { op_type: 1, row }],
         }
     }
 
     fn delete_op(table_id: TableId, table_name: &str, row: ProductValue) -> DatabaseTableUpdate {
-        let row_pk = row.to_data_key().to_bytes();
         DatabaseTableUpdate {
             table_id,
             table_name: table_name.to_string(),
-            ops: vec![TableOp {
-                op_type: 0,
-                row,
-                row_pk,
-            }],
+            ops: vec![TableOp { op_type: 0, row }],
         }
     }
 
@@ -271,7 +260,6 @@ mod tests {
 
         let op = TableOp {
             op_type: 1,
-            row_pk: row.to_data_key().to_bytes(),
             row: row.clone(),
         };
 
@@ -405,46 +393,6 @@ mod tests {
     }
 
     #[test]
-    fn test_eval_incr_maintains_row_ids() -> ResultTest<()> {
-        let (db, _) = make_test_db()?;
-        let mut tx = db.begin_mut_tx(IsolationLevel::Serializable);
-
-        let schema = ProductType::from([("u8", AlgebraicType::U8)]);
-        let row = product!(1u8);
-
-        // generate row id from row
-        let id1 = &row.to_data_key().to_bytes();
-
-        // create table empty table "test"
-        let table_id = create_table_with_rows(&db, &mut tx, "test", schema.clone(), &[])?;
-
-        // select * from test
-        let query: ExecutionSet = QueryExpr::new(db_table(schema.clone(), table_id)).try_into()?;
-
-        let op = TableOp {
-            op_type: 0,
-            row_pk: id1.clone(),
-            row: row.clone(),
-        };
-
-        let update = DatabaseTableUpdate {
-            table_id,
-            table_name: "test".into(),
-            ops: vec![op],
-        };
-
-        let update = DatabaseUpdate { tables: vec![update] };
-        db.rollback_mut_tx(&ExecutionContext::default(), tx);
-        let tx = db.begin_tx();
-        let result = query.eval_incr(&db, &tx, &update, AuthCtx::for_testing())?;
-        let id2 = &result.tables[0].ops[0].row_pk;
-
-        // check that both row ids are the same
-        assert_eq!(id1, id2);
-        Ok(())
-    }
-
-    #[test]
     fn test_eval_incr_for_index_scan() -> ResultTest<()> {
         let (db, _tmp) = make_test_db()?;
 
@@ -460,12 +408,7 @@ mod tests {
             db.insert(&mut tx, table_id, row)?;
 
             let row = product!(i + 10, i);
-            let row_pk = row.to_data_key().to_bytes();
-            ops.push(TableOp {
-                op_type: 0,
-                row_pk,
-                row,
-            })
+            ops.push(TableOp { op_type: 0, row })
         }
 
         let update = DatabaseUpdate {
@@ -500,7 +443,6 @@ mod tests {
 
         assert_eq!(op.op_type, 0);
         assert_eq!(op.row, product!(13u64, 3u64));
-        assert_eq!(op.row_pk, product!(13u64, 3u64).to_data_key().to_bytes());
         Ok(())
     }
 
@@ -832,7 +774,6 @@ mod tests {
 
         let row2 = TableOp {
             op_type: 1,
-            row_pk: row.to_data_key().to_bytes(),
             row: row.clone(),
         };
 
@@ -952,17 +893,9 @@ mod tests {
         let s = compile_read_only_query(&db, &tx, &AuthCtx::for_testing(), SUBSCRIBE_TO_ALL_QUERY)?.into();
         check_query_eval(&db, &tx, &s, 2, &[row_1.clone(), row_2.clone()])?;
 
-        let row1 = TableOp {
-            op_type: 0,
-            row_pk: row_1.to_data_key().to_bytes(),
-            row: row_1,
-        };
+        let row1 = TableOp { op_type: 0, row: row_1 };
 
-        let row2 = TableOp {
-            op_type: 1,
-            row_pk: row_2.to_data_key().to_bytes(),
-            row: row_2,
-        };
+        let row2 = TableOp { op_type: 1, row: row_2 };
 
         let data1 = DatabaseTableUpdate {
             table_id: schema_1.table_id,