Closes #831.
Closes #874.
Closes #280.

```
full-scan               time:   [105.90 ms 106.18 ms 106.51 ms]
                        change: [-53.505% -53.317% -53.150%] (p = 0.00 < 0.05)
                        Performance has improved.

full-join               time:   [312.20 µs 313.41 µs 314.63 µs]
                        change: [-33.453% -33.176% -32.873%] (p = 0.00 < 0.05)
                        Performance has improved.

incr-select             time:   [2.0535 µs 2.0584 µs 2.0649 µs]
                        change: [-12.335% -12.142% -11.936%] (p = 0.00 < 0.05)
                        Performance has improved.

incr-join               time:   [10.726 µs 10.747 µs 10.769 µs]
                        change: [+0.3173% +0.5243% +0.7765%] (p = 0.00 < 0.05)
                        Change within noise threshold.
```

Removes some `#[tracing::instrument(skip_all)]` as well.

# Expected complexity level and risk

1